:bricks: Introduce Cloudflare worker as reverse proxy
The Cloud Run certificate would expire every ~3 months leading to the
custom domain mapping failing with an "Invalid SSL certificate" error
and status code 526.
In the new setup the domain is mapped to the Cloudflare worker which
acts as a reverse proxy and routes the requests to the Cloud Run
generated domain.

Note we had to edit the user API token config in the UI to give it
"Workers Routes:Edit" & "Workers Scripts:Edit" permissions, refs:
https://dash.cloudflare.com/profile/api-tokens

The error was:
```
Error: error creating worker script: Authentication error (10000)
  with cloudflare_worker_script.grafana_reverse_proxy,
  on cloudflare.tf line 56, in resource "cloudflare_worker_script" "grafana_reverse_proxy":
  56: resource "cloudflare_worker_script" "grafana_reverse_proxy" {
```

On the pricing side, Cloudflare offers 100k free requests per month
across all workers or 10M requests for $5 a month.